Create Session 10
Today's session was very difficult for a number of reasons. Mainly to do with some personal mental health problems that kept me quite distracted. Not to mention that the complexity of the challenges jumped. Overall, a tough session, but nonetheless highly educational as a I learned quite a bit from it, hence why I only managed to solve two challenges.